refactor: 대학 선택 관련 함수를 커스텀 훅으로 분리
